feat: Separate copyright model
